+++ src/_/url.ts
@@ -5,33 +5,26 @@ const FULL_PATH = 4;
 const MEDIA = 2;
 const QUERY = 6;
 const RESERVED = 5;
-const SLASHES = 3;
 const SPECIAL = 1;
 const VERSION = `v1`;
 
-export const createUrl = (
-    { debug, domain, path, output, quality, src, transform }: CreateUrlData
-): string => {
+export const isAbsolute = ( src: string, domain: string ) => ( src.slice( 0, domain.length + 1 ) === `${ domain }/` );
 
-    const _domain = `${ domain }/`;
-    const isAbsolute = ( src.slice( 0, _domain.length ) === _domain );
-    const _path = isAbsolute ? `media:${ src.slice( _domain.length ) }` : src;
-    const parsed = rPath.exec( _path );
+export const createUrl = (
+    { debug, domain, output, quality, src, transform }: CreateUrlData
+): string => {
+    const _isAbsolute = isAbsolute( src, domain );
+    const path = _isAbsolute ? `media:${ src.slice( `${ domain }/`.length ) }` : src;
+    const parsed = rPath.exec( path );
     const isMedia = parsed && parsed[ MEDIA ];
     const actualDebug = debug ? `/debug` : ``;
     const actualOutput = output ? `/output=${ output }` : ``;
-    // eslint-disable-next-line no-nested-ternary
-    const actualPath = isMedia ? (
-        ( path && !isAbsolute ) ?
-            `${ path }${ parsed[ SLASHES ] ? parsed[ SLASHES ][ 0 ] : `` }${ parsed[ FULL_PATH ] }` :
-            parsed[ FULL_PATH ]
-    ) : _path;
+    const actualPath = isMedia ? parsed[ FULL_PATH ] : path;
     const actualQuality = quality ? `/quality=${ quality }` : ``;
     const actualTransform = transform ? `/${ transform }` : ``;
-
     return `${
-        _domain
-    }${
+        domain
+    }/${
         ( parsed && ( parsed[ SPECIAL ] || parsed[ RESERVED ] ) ) ?
         // old syntax
         `${