+++ src/next/placeholder.tsx
@@ -1,23 +1,21 @@
+/* eslint-disable no-undef */
 import { createUrl } from '../_/url';
+import { parseSrc, getDomain } from './utils';
 
 const PLACEHOLDER_SIZE = 1000;
 export default async ( src: string ): Promise< string > => {
     const response = await fetch( createUrl(
         {
-            // eslint-disable-next-line no-undef
-            "domain": process.env.NEXT_PUBLIC_TWICPICS_DOMAIN,
-            // eslint-disable-next-line no-undef
-            "path": process.env.NEXT_PUBLIC_TWICPICS_PATH,
-            src,
-            "transform": `resize=${ PLACEHOLDER_SIZE }`,
+            "domain": getDomain(),
             "output": `preview`,
+            "src": parseSrc( src ),
+            "transform": `resize=${ PLACEHOLDER_SIZE }`,
         }
     ) );
     const arrayBuffer = await response.arrayBuffer();
     const dataURL = `data:${
         response.headers.get( `content-type` )
     };base64,${
-        // eslint-disable-next-line no-undef
         Buffer.from( arrayBuffer ).toString( `base64` )
     }`;
     return dataURL;